+++ source/TEST/SuffixArray_test.C
@@ -69,7 +69,7 @@ CHECK((virtual String toString()=0))
 }
 RESULT
 
-CHECK((virtual void findSpec(std::vector< std::vector< std::pair< std::pair< int, int >, float > > > &candidates, const std::vector< double > &spec)=0))
+CHECK((virtual void findSpec(std::vector< std::vector< std::pair< std::pair< int, int >, double > > > &candidates, const std::vector< double > &spec)=0))
 {
   NOT_TESTABLE
 }
